virtio-ramfb: expose framebuffer VRAM in BAR
Windows requires that the framebuffer be defined as a resource list entry.
For PCIe devices, this means that the FB address should be in one of the
device's BAR. ramfb currently waits for the UEFI driver to set the
resolution before allocating memory and passing the address of that
allocation to QEMU via fw_config. We used the same code for virtio-ramfb
before but now, we break away from ramfb (standalone) with a slightly
different (and incompatible) implementation where the framebuffer address
is no longer determined by the driver and is fixed by the device.